pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitmemory"
version = "0.1.0"
description = "Git-aware memory system for AI coding assistants"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "gitpython>=3.1.0",
    "textual>=0.40.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "tree-sitter-languages>=1.10.0",
    "tree-sitter>=0.21.0,<0.22",  # Pin for tree-sitter-languages compatibility
]

[project.scripts]
view-fn-hist = "view_fn_hist.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/view_fn_hist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
